// src/lib/huawei/services/dcs.types.ts
export type DcsRedisInstance = {
  capacity: string;
  createdAt: string;
  engine: string;
  engineVersion: string;
  id: string;
  ip: string;
  mode: string;
  name: string;
  port: string;
  projectId: string;
  projectName: string;
  region: string;
  status: string;
  usedMemory: string;
  vpcId: string;
};